requirements.txt
coverage==4.2
dj-database-url==0.4.0
Django==1.9.2
django-admin-honeypot==1.0.0
django-filter==1.0.1
djangorestframework==3.5.4
gunicorn==19.4.5
Markdown==2.6.8
Pillow==3.3.1
psycopg2==2.6.1
whitenoise==2.0.6
social-auth-core==1.2.0
social-auth-app-django==1.1.0
boto3==1.4.4
django-storages==1.5.2
boto==2.46.1
django-letsencrypt==1.1.0
